feat(local-api): D1 — extract routes/channels.py (closes #964)
