+++ scripts/local_api/__init__.py
@@ -1,0 +1,1 @@
+"""Local API package namespace for extracted route modules."""
