+++ scripts/local_api/routes/__init__.py
@@ -1,0 +1,1 @@
+"""Route modules for the local API."""
